tools/requirements.txt
nats-py>=2.3.1
asyncio
python-dateutil
requests
